requirements.txt
asgiref==3.7.2 ; python_version >= "3.9" and python_version < "4.0"
django-cors-headers==4.2.0 ; python_version >= "3.9" and python_version < "4.0"
django-markdownx==4.0.2 ; python_version >= "3.9" and python_version < "4.0"
django==4.2.5 ; python_version >= "3.9" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.9" and python_version < "4.0"
importlib-metadata==6.8.0 ; python_version >= "3.9" and python_version < "3.10"
markdown==3.4.4 ; python_version >= "3.9" and python_version < "4.0"
pillow==10.0.1 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.9" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
zipp==3.16.2 ; python_version >= "3.9" and python_version < "3.10"
